diff3: fix race condition involving directories
This fixes a TOC-TOU race with "diff3 A B C" when one
of its arguments is a directory.
* src/diff.c (no_directory): New static var.
(NO_DIRECTORY_OPTION): New constant.
(longopts): Add ---no-directory.
(main): Support it.
(compare_files): If ---no-directory is given, do not treat
directories specially.
* src/diff3.c (main): Do not check whether an argument is
a directory.
(read_diff): Pass ---no-directory to diff, so that it checks.