cmp: work around lseek bugs
* src/cmp.c (ignore_initial): Now an array of intmax_t not off_t.
This allows cmp -i N to work better on platforms where intmax_t
is wider than off_t, N fits in the former and not the latter,
and the input is not a regular file.  All uses changed.
(cmp): Instead of seeking to file end when -i N is huge,
simply pretend that we read zero bytes.  This avoids some
kernel lseek bugs.