diff3: simplify by assuming !GCC_LINT
GCC's static checks no longer seem to warn about the code
protected by GCC_LINT, so remove its last uses, which were in diff3.
(Or perhaps it was valgrind or gcc -fsanitize=address leak
checking instead?  But in that case there are other false
positives and we don’t worry about them either.)  This code was
problematic anyway when introduced (as witness the corresponding
test case) and it slowed diff3 down a bit.
* configure.ac (GCC_LINT): Remove.  All uses removed.
* src/diff3.c: Simplify by assuming !GCC_LINT.